+++ src/Model/Pdf.php
@@ -27,7 +27,7 @@ declare(strict_types=1);
 
 namespace Vianetz\Pdf\Model;
 
-use Vianetz\Pdf\Exception;
+use Vianetz\Pdf\InvalidDocumentException;
 use Vianetz\Pdf\Model\Generator\AbstractGenerator;
 use Vianetz\Pdf\NoDataException;
 
@@ -95,6 +95,8 @@ class Pdf implements CanSave, Pdfable
     final public function attach(string $file): self
     {
         $this->attachments[] = $file;
+        // Reset cached pdf contents.
+        $this->pdfContents = null;
 
         return $this;
     }
@@ -127,7 +129,7 @@ class Pdf implements CanSave, Pdfable
     /**
      * Return merged pdf contents of all documents and save it to single temporary files.
      *
-     * @throws \Vianetz\Pdf\NoDataException|\Vianetz\Pdf\Exception
+     * @throws \Vianetz\Pdf\NoDataException|\Vianetz\Pdf\InvalidDocumentException
      */
     private function renderPdfContentsForAllDocuments(): void
     {
@@ -145,8 +147,11 @@ class Pdf implements CanSave, Pdfable
                 }
             } elseif ($documentInstance instanceof Pdfable) {
                 $pdfContents = $documentInstance->toPdf();
+                if (empty($pdfContents)) {
+                    continue;
+                }
             } else {
-                throw new Exception('invalid document type');
+                throw new InvalidDocumentException('invalid document type');
             }
 
             if ($documentInstance instanceof HasBackgroundPdf) {